Split audio workspace helpers
