+++ frontend/app/(core)/(workspace)/app/audio/AudioWorkspace.tsx
@@ -7,23 +7,19 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import {
   AudioLines,
-  Check,
   CheckCircle2,
   ChevronDown,
-  Clapperboard,
   Clock3,
   Coins,
   FileVideo,
   Gauge,
   Languages,
   Mic2,
-  Music2,
   Play,
   SlidersHorizontal,
   Sparkles,
   Upload,
   Video,
-  type LucideIcon,
 } from 'lucide-react';
 
 import { useRequireAuth } from '@/hooks/useRequireAuth';
@@ -33,7 +29,6 @@ import { useI18n } from '@/lib/i18n/I18nProvider';
 import type { Job } from '@/types/jobs';
 import { Button, ButtonLink } from '@/components/ui/Button';
 import { Textarea } from '@/components/ui/Input';
-import { SelectMenu } from '@/components/ui/SelectMenu';
 import { UIIcon } from '@/components/ui/UIIcon';
 import {
   AUDIO_INTENSITY_VALUES,
@@ -65,377 +60,39 @@ import {
   type AudioVoiceProfile,
 } from '@/lib/audio-generation';
 import AudioLatestRendersRail from './AudioLatestRendersRail';
+import { AudioModePicker, AudioSelectControl, ToggleRow } from './_components/audio-workspace-controls';
+import {
+  AUDIO_VOICE_GENDER_VALUES,
+  DEFAULT_INTENSITY,
+  DEFAULT_LANGUAGE,
+  DEFAULT_MANUAL_DURATION_SEC,
+  DEFAULT_MOOD,
+  DEFAULT_PACK,
+  DEFAULT_VOICE_DELIVERY,
+  DEFAULT_VOICE_GENDER,
+  DEFAULT_VOICE_PROFILE,
+  fetchJobDetail,
+  formatCopy,
+  formatCurrency,
+  formatDateTime,
+  inferOutputKind,
+  probeVideoDuration,
+  resolveProviderLabel,
+  resolveUiErrorMessage,
+  uploadAsset,
+} from './_lib/audio-workspace-helpers';
+import type {
+  ActiveAudioJobState,
+  AudioResultState,
+  GeneratedSourceVideo,
+  SourceVideoState,
+} from './_lib/audio-workspace-types';
 import {
   buildAudioModeOptions,
   DEFAULT_AUDIO_WORKSPACE_COPY,
   formatAudioOutputKind,
   type AudioWorkspaceCopy,
 } from './copy';
-
-type SourceVideoState = {
-  url: string;
-  jobId?: string | null;
-  thumbUrl?: string | null;
-  durationSec?: number | null;
-  aspectRatio?: string | null;
-  label: string;
-};
-
-type GeneratedSourceVideo = {
-  jobId: string;
-  url: string;
-  thumbUrl: string | null;
-  durationSec: number | null;
-  aspectRatio: string | null;
-  label: string;
-  createdAt: string;
-  hasAudio: boolean;
-};
-
-type AudioJobSettingsSnapshot = {
-  pack?: string | null;
-  prompt?: string | null;
-  mood?: string | null;
-  intensity?: string | null;
-  durationSec?: number | null;
-  script?: string | null;
-  musicEnabled?: boolean | null;
-  exportAudioFile?: boolean | null;
-  voiceGender?: string | null;
-  voiceProfile?: string | null;
-  voiceDelivery?: string | null;
-  language?: string | null;
-  outputKind?: AudioOutputKind | null;
-  sourceJobId?: string | null;
-  sourceVideoUrl?: string | null;
-  refs?: {
-    sourceVideoUrl?: string | null;
-    voiceSampleUrl?: string | null;
-  } | null;
-};
-
-type AudioJobDetail = {
-  ok?: boolean;
-  error?: string;
-  jobId: string;
-  surface?: string;
-  status?: 'pending' | 'running' | 'completed' | 'failed' | null;
-  progress?: number | null;
-  message?: string | null;
-  videoUrl?: string | null;
-  audioUrl?: string | null;
-  thumbUrl?: string | null;
-  aspectRatio?: string | null;
-  engineLabel?: string | null;
-  durationSec?: number | null;
-  settingsSnapshot?: AudioJobSettingsSnapshot | null;
-};
-
-type ActiveAudioJobState = {
-  jobId: string;
-  status: 'pending' | 'running' | 'completed' | 'failed';
-  progress: number;
-  message: string | null;
-  videoUrl: string | null;
-  audioUrl: string | null;
-  thumbUrl: string | null;
-  outputKind: AudioOutputKind;
-};
-
-type AudioResultState = {
-  jobId: string;
-  videoUrl: string | null;
-  audioUrl: string | null;
-  thumbUrl: string | null;
-  outputKind: AudioOutputKind;
-};
-
-const DEFAULT_PACK: AudioPackId = 'cinematic';
-const DEFAULT_MOOD: AudioMood = 'epic';
-const DEFAULT_INTENSITY: AudioIntensity = 'standard';
-const DEFAULT_VOICE_GENDER: AudioVoiceGender = 'female';
-const DEFAULT_VOICE_PROFILE: AudioVoiceProfile = 'balanced';
-const DEFAULT_VOICE_DELIVERY: AudioVoiceDelivery = 'cinematic';
-const DEFAULT_LANGUAGE: AudioLanguage = 'auto';
-const DEFAULT_MANUAL_DURATION_SEC = 8;
-const AUDIO_VOICE_GENDER_VALUES = ['female', 'male', 'neutral'] as const;
-const AUDIO_MODE_META: Record<
-  AudioPackId,
-  {
-    icon: LucideIcon;
-    providerKey: keyof AudioWorkspaceCopy['controls']['providers'];
-  }
-> = {
-  music_only: {
-    icon: Music2,
-    providerKey: 'music',
-  },
-  voice_only: {
-    icon: Mic2,
-    providerKey: 'voice',
-  },
-  cinematic: {
-    icon: Clapperboard,
-    providerKey: 'sfx',
-  },
-  cinematic_voice: {
-    icon: AudioLines,
-    providerKey: 'mix',
-  },
-};
-
-function resolveProviderLabel(copy: AudioWorkspaceCopy, pack: AudioPackId): string {
-  return copy.controls.providers[AUDIO_MODE_META[pack].providerKey];
-}
-
-function formatDateTime(value: string, locale?: string): string {
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) return value;
-  return new Intl.DateTimeFormat(locale, {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  }).format(parsed);
-}
-
-function formatCopy(template: string, values: Record<string, string | number>): string {
-  return Object.entries(values).reduce((resolved, [key, value]) => {
-    return resolved.split(`{${key}}`).join(String(value));
-  }, template);
-}
-
-function formatCurrency(amountCents?: number | null, currency = 'USD', locale?: string): string {
-  if (typeof amountCents !== 'number') return '-';
-  try {
-    return new Intl.NumberFormat(locale, {
-      style: 'currency',
-      currency,
-    }).format(amountCents / 100);
-  } catch {
-    return `${currency} ${(amountCents / 100).toFixed(2)}`;
-  }
-}
-
-function resolveUiErrorMessage(
-  error: unknown,
-  fallback: string,
-  genericMessages: string[] = []
-): string {
-  if (!(error instanceof Error) || !error.message.trim().length) {
-    return fallback;
-  }
-  return genericMessages.includes(error.message) ? fallback : error.message;
-}
-
-function inferOutputKind(detail: {
-  outputKind?: AudioOutputKind | null;
-  videoUrl?: string | null;
-  audioUrl?: string | null;
-}): AudioOutputKind {
-  if (detail.outputKind) return detail.outputKind;
-  if (detail.videoUrl && detail.audioUrl) return 'both';
-  if (detail.audioUrl) return 'audio';
-  return 'video';
-}
-
-async function probeVideoDuration(url: string): Promise<number | null> {
-  return new Promise((resolve) => {
-    const video = document.createElement('video');
-    const cleanup = () => {
-      video.pause();
-      video.removeAttribute('src');
-      video.load();
-    };
-    const finish = (value: number | null) => {
-      cleanup();
-      resolve(value);
-    };
-
-    video.preload = 'metadata';
-    video.muted = true;
-    video.onloadedmetadata = () => {
-      finish(Number.isFinite(video.duration) ? Math.round(video.duration) : null);
-    };
-    video.onerror = () => finish(null);
-    video.src = url;
-  });
-}
-
-async function uploadAsset(file: File, kind: 'video' | 'audio'): Promise<{ url: string; name: string }> {
-  const formData = new FormData();
-  formData.append('file', file, file.name);
-  const response = await authFetch(kind === 'video' ? '/api/uploads/video' : '/api/uploads/audio', {
-    method: 'POST',
-    body: formData,
-  });
-  const payload = (await response.json().catch(() => null)) as
-    | { ok?: boolean; error?: string; asset?: { url?: string; name?: string } }
-    | null;
-  if (!response.ok || !payload?.ok || !payload.asset?.url) {
-    throw new Error(payload?.error ?? 'Upload failed');
-  }
-  return {
-    url: payload.asset.url,
-    name: payload.asset.name ?? file.name,
-  };
-}
-
-async function fetchJobDetail(jobId: string): Promise<AudioJobDetail> {
-  const response = await authFetch(`/api/jobs/${encodeURIComponent(jobId)}`);
-  const payload = (await response.json().catch(() => null)) as AudioJobDetail | null;
-  if (!response.ok || !payload?.jobId) {
-    throw new Error(payload?.error ?? 'Unable to load job.');
-  }
-  return payload;
-}
-
-function AudioModePicker({
-  value,
-  options,
-  onChange,
-}: {
-  value: AudioPackId;
-  options: Array<{ id: AudioPackId; label: string; description: string }>;
-  onChange: (pack: AudioPackId) => void;
-}) {
-  return (
-    <div className="grid gap-2 md:grid-cols-2 2xl:grid-cols-4">
-      {options.map((mode) => (
-        <AudioModeCard
-          key={mode.id}
-          mode={mode}
-          active={value === mode.id}
-          onClick={() => onChange(mode.id)}
-        />
-      ))}
-    </div>
-  );
-}
-
-function AudioModeCard({
-  mode,
-  active,
-  onClick,
-}: {
-  mode: { id: AudioPackId; label: string; description: string };
-  active: boolean;
-  onClick: () => void;
-}) {
-  const Icon = AUDIO_MODE_META[mode.id].icon;
-
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      aria-pressed={active}
-      className={[
-        'group relative min-h-[86px] overflow-hidden rounded-[10px] border px-3.5 py-3 text-left transition duration-200',
-        active
-          ? 'border-brand bg-brand-soft shadow-[0_14px_34px_rgba(46,99,216,0.12)] ring-1 ring-brand/25'
-          : 'border-hairline bg-surface hover:border-brand/45 hover:bg-surface-hover hover:shadow-card',
-      ].join(' ')}
-    >
-      <div className="flex items-start gap-3">
-        <span
-          className={[
-            'flex h-10 w-10 shrink-0 items-center justify-center rounded-[9px] transition',
-            active
-              ? 'bg-brand text-on-brand shadow-[0_10px_28px_rgba(46,99,216,0.22)]'
-              : 'bg-brand-soft text-brand group-hover:bg-brand-soft-hover',
-          ].join(' ')}
-        >
-          <UIIcon icon={Icon} size={21} strokeWidth={1.9} />
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="block text-sm font-semibold text-text-primary">{mode.label}</span>
-          <span className="mt-1 block text-xs leading-5 text-text-secondary">{mode.description}</span>
-        </span>
-        <span
-          className={[
-            'flex h-6 w-6 shrink-0 items-center justify-center rounded-full border transition',
-            active
-              ? 'border-brand bg-brand text-on-brand'
-              : 'border-hairline bg-surface text-transparent group-hover:border-brand/45',
-          ].join(' ')}
-          aria-hidden
-        >
-          {active ? <Check className="h-3.5 w-3.5" /> : null}
-        </span>
-      </div>
-    </button>
-  );
-}
-
-function AudioInlineSelectLabel({ icon, label }: { icon: LucideIcon; label: string }) {
-  return (
-    <span className="inline-flex min-w-0 items-center gap-2">
-      <UIIcon icon={icon} size={15} strokeWidth={1.8} />
-      <span className="truncate">{label}</span>
-    </span>
-  );
-}
-
-function AudioSelectControl({
-  label,
-  value,
-  options,
-  icon,
-  onChange,
-}: {
-  label: string;
-  value: string | number | boolean;
-  options: Array<{ value: string | number | boolean; label: string; disabled?: boolean }>;
-  icon: LucideIcon;
-  onChange: (next: string | number | boolean) => void;
-}) {
-  return (
-    <div className="min-w-0">
-      <span className="mb-2 block text-sm font-semibold text-text-primary">{label}</span>
-      <SelectMenu
-        options={options.map((option) => ({
-          ...option,
-          label: <AudioInlineSelectLabel icon={icon} label={option.label} />,
-        }))}
-        value={value}
-        onChange={(next) => onChange(String(next))}
-        className="min-w-0"
-        buttonClassName="min-h-0 h-10 rounded-full border-border bg-surface px-3 py-0 text-[12px] font-medium shadow-none dark:border-white/10 dark:bg-surface dark:text-white/92 dark:hover:border-white/16 dark:hover:bg-surface-hover"
-        menuClassName="min-w-[180px]"
-        menuPlacement="top"
-      />
-    </div>
-  );
-}
-
-function ToggleRow({
-  label,
-  description,
-  checked,
-  onChange,
-}: {
-  label: string;
-  description: string;
-  checked: boolean;
-  onChange: (next: boolean) => void;
-}) {
-  return (
-    <label className="flex items-start justify-between gap-4 rounded-[10px] border border-hairline bg-surface px-4 py-3 shadow-sm">
-      <div className="min-w-0">
-        <p className="text-sm font-semibold text-text-primary">{label}</p>
-        {description ? <p className="mt-1 text-sm text-text-secondary">{description}</p> : null}
-      </div>
-      <span className="relative mt-1 inline-flex h-6 w-11 shrink-0 items-center">
-        <input
-          type="checkbox"
-          checked={checked}
-          onChange={(event) => onChange(event.target.checked)}
-          className="peer sr-only"
-        />
-        <span className="absolute inset-0 rounded-full bg-border transition peer-checked:bg-brand" />
-        <span className="absolute left-0.5 top-0.5 h-5 w-5 rounded-full bg-white shadow transition peer-checked:translate-x-5" />
-      </span>
-    </label>
-  );
-}
 
 export default function AudioWorkspace() {
   const searchParams = useSearchParams();
